feat: Create local storage handler w/ custom hook
Added a custom hook to handle the localStorage.

The functionality is very easy, you create an instance of the
[UseLocalStorage] with the proper props and you'r done!!!

By default it saves a prefixed version of the given [key]
('chatappclone'), this happens to prevent an unwanted override.

The function could receive a function that will be executed and then
saved as a JSON or a variable that will parse and save.

For more information read the documentation here: https://reactjs.org/docs/hooks-custom.html